requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
exceptiongroup==1.1.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
python-dotenv==1.0.0
python-telegram-bot==20.4
requests==2.31.0
schedule==1.2.0
sniffio==1.3.0
soupsieve==2.4.1
tomli==2.0.1
urllib3==2.0.3
yarl==1.9.2
